+++ src/components/common/Input.tsx
@@ -10,8 +10,8 @@ import "react-date-range/dist/styles.css";
 import "react-date-range/dist/theme/default.css";
 
 interface Option {
+  id: number;
   label: string;
-  value: string;
 }
 
 interface InputProps {
@@ -101,7 +101,7 @@ const Input = (props: InputProps) => {
             disabled={disabled}
           >
             {options.map((option) => (
-              <option key={option.value} value={option.value}>
+              <option key={option.id} value={option.id}>
                 {option.label}
               </option>
             ))}
