+++ src/components/common/index.ts
@@ -1,0 +1,8 @@
+export { default as Button } from "./Button";
+export { default as Checkbox } from "./Checkbox";
+export { default as Dropdown } from "./Dropdown";
+export { default as Header } from "./Header";
+export { default as Input } from "./Input";
+export { default as Label } from "./Label";
+export { default as Pagination } from "./Pagination";
+export { default as Title } from "./Title";
